+++ packages/jsActions/nanoflow-actions-native/src/geolocation/GetCurrentLocationMinimumAccuracy.ts
@@ -8,15 +8,19 @@ import { Big } from "big.js";
 import Geolocation, {
     GeolocationError,
     GeolocationOptions,
-    GeolocationResponse
+    GeolocationResponse,
+    GeolocationStatic
 } from "@react-native-community/geolocation";
+
+import type { Platform, NativeModules } from "react-native";
+import type { GeoError, GeoPosition, GeoOptions } from "../../typings/Geolocation";
 
 /**
  * This action retrieves the current geographical position of a user/device with a minimum accuracy as parameter. If a position is not acquired with minimum accuracy within a specific timeout it will retrieve the last most precise location.
  *
  * Since this can compromise privacy, the position is not available unless the user approves it. The web browser will request the permission at the first time the location is requested. When denied by the user it will not prompt a second time.
  *
- * On hybrid and native platforms the permission can be requested with the `RequestLocationPermission` action.
+ * On hybrid and native platforms the permission should be requested with the `RequestLocationPermission` action.
  *
  * Best practices:
  * https://developers.google.com/web/fundamentals/native-hardware/user-location/
@@ -34,17 +38,44 @@ export async function GetCurrentLocationMinimumAccuracy(
 ): Promise<mendix.lib.MxObject> {
     // BEGIN USER CODE
 
-    if (navigator && navigator.product === "ReactNative" && !navigator.geolocation) {
-        (navigator.geolocation as any) = Geolocation;
+    let reactNativeModule: { NativeModules: typeof NativeModules; Platform: typeof Platform } | undefined;
+    let geolocationModule:
+        | Geolocation
+        | GeolocationStatic
+        | typeof import("react-native-geolocation-service")
+        | undefined;
+
+    if (navigator && navigator.product === "ReactNative") {
+        reactNativeModule = require("react-native");
+
+        if (!reactNativeModule) {
+            return Promise.reject(new Error("React Native module could not be found"));
+        }
+
+        if (reactNativeModule.NativeModules.RNFusedLocation) {
+            geolocationModule = (await import("react-native-geolocation-service")).default;
+        } else if (reactNativeModule.NativeModules.RNCGeolocation) {
+            geolocationModule = Geolocation;
+        } else {
+            return Promise.reject(new Error("Geolocation module could not be found"));
+        }
+    } else if (navigator && navigator.geolocation) {
+        geolocationModule = navigator.geolocation;
+    } else {
+        return Promise.reject(new Error("Geolocation module could not be found"));
     }
 
     return new Promise((resolve, reject) => {
+        if (!geolocationModule) {
+            return reject(new Error("Geolocation module could not be found"));
+        }
+
         const options = getOptions();
 
         // This action is only required while running in PWA or hybrid.
         if (navigator && (!navigator.product || navigator.product !== "ReactNative")) {
             // This ensures the browser will not ignore the maximumAge https://stackoverflow.com/questions/3397585/navigator-geolocation-getcurrentposition-sometimes-works-sometimes-doesnt/31916631#31916631
-            navigator.geolocation.getCurrentPosition(
+            geolocationModule.getCurrentPosition(
                 // eslint-disable-next-line @typescript-eslint/no-empty-function
                 () => {},
                 // eslint-disable-next-line @typescript-eslint/no-empty-function
@@ -52,11 +83,12 @@ export async function GetCurrentLocationMinimumAccuracy(
                 {}
             );
         }
-        const watchId: number = navigator.geolocation.watchPosition(onSuccess, onError, options);
-        const timeStart = Date.now();
-        let lastAccruedPosition: GeolocationResponse;
 
-        function createGeolocationObject(position: GeolocationResponse): void {
+        const watchId: number = geolocationModule.watchPosition(onSuccess, onError, options);
+        const timeStart = Date.now();
+        let lastAccruedPosition: GeolocationResponse | GeoPosition;
+
+        function createGeolocationObject(position: GeolocationResponse | GeoPosition): void {
             mx.data.create({
                 entity: "NanoflowCommons.Geolocation",
                 callback: mxObject => resolve(mapPositionToMxObject(mxObject, position)),
@@ -65,9 +97,9 @@ export async function GetCurrentLocationMinimumAccuracy(
             });
         }
 
-        function onSuccess(position: GeolocationResponse): void {
+        function onSuccess(position: GeolocationResponse | GeoPosition): void {
             if (watchId && (!minimumAccuracy || minimumAccuracy >= position.coords.accuracy)) {
-                navigator.geolocation.clearWatch(watchId);
+                geolocationModule?.clearWatch(watchId);
                 createGeolocationObject(position);
             } else {
                 if (!lastAccruedPosition || position.coords.accuracy < lastAccruedPosition.coords.accuracy) {
@@ -75,19 +107,30 @@ export async function GetCurrentLocationMinimumAccuracy(
                 }
                 const timeDiff = Date.now() - timeStart;
                 if (!timeout || timeout.lte(timeDiff)) {
-                    navigator.geolocation.clearWatch(watchId);
+                    geolocationModule?.clearWatch(watchId);
                     createGeolocationObject(lastAccruedPosition);
                 }
             }
         }
 
-        function onError(error: GeolocationError): void {
+        function onError(error: GeolocationError | GeoError): void {
             return reject(new Error(error.message));
         }
 
-        function getOptions(): GeolocationOptions {
-            const timeoutNumber = timeout && Number(timeout.toString());
+        function getOptions(): GeolocationOptions | GeoOptions {
+            let timeoutNumber = timeout && Number(timeout.toString());
             const maximumAgeNumber = maximumAge && Number(maximumAge.toString());
+
+            // If the timeout is 0 or undefined (empty), it causes a crash on iOS.
+            // If the timeout is undefined (empty); we set timeout to 30 sec (default timeout)
+            // If the timeout is 0; we set timeout to 1 hour (no timeout)
+            if (reactNativeModule?.Platform.OS === "ios") {
+                if (timeoutNumber === undefined) {
+                    timeoutNumber = 30000;
+                } else if (timeoutNumber === 0) {
+                    timeoutNumber = 3600000;
+                }
+            }
 
             return {
                 timeout: timeoutNumber,
@@ -98,7 +141,7 @@ export async function GetCurrentLocationMinimumAccuracy(
 
         function mapPositionToMxObject(
             mxObject: mendix.lib.MxObject,
-            position: GeolocationResponse
+            position: GeolocationResponse | GeoPosition
         ): mendix.lib.MxObject {
             mxObject.set("Timestamp", new Date(position.timestamp));
             mxObject.set("Latitude", new Big(position.coords.latitude.toFixed(8)));